krummserver/requirements.txt
django-cors-headers >= 2.1.0
